flujo de creacion de salas de un estudio (crear/editar/eliminar)
