+++ app/createRoom/page.tsx
@@ -1,0 +1,11 @@
+
+"use client";
+
+import CrearSalaForm from "@/components/createRoom/CrearSalaForm"; 
+export default function CrearSalaPage() {
+  return (
+    <div>
+      <CrearSalaForm />
+    </div>
+  );
+}
