+++ package.json
@@ -18,6 +18,8 @@
     "react-dom": "19.1.0",
     "react-icons": "^5.5.0",
     "react-leaflet": "^5.0.0",
+    "react-toastify": "^11.0.5",
+    "sweetalert2": "^11.23.0",
     "yup": "^1.7.0",
     "zustand": "^5.0.8"
   },
