+++ services/instruments.service.ts
@@ -1,0 +1,20 @@
+import { http } from "@/lib/Http";
+
+export const instrumentsService = {
+  createInstrument: async (instrument: {
+    name: string;
+    description: string;
+    price: number;
+    available: boolean;
+    categoryName: string;
+    studioId: string;
+  }) => {
+    try {
+      const response = await http.post("/instruments/create", instrument);
+      return response.data; 
+    } catch (error: any) {
+      console.error("Error en instrumentsService.createInstrument:", error);
+      throw new Error("No se pudo crear el instrumento");
+    }
+  },
+}
