+++ services/registerStudio.services.ts
@@ -1,4 +1,4 @@
-// src/services/studio.services.ts
+
 import { http } from "@/lib/Http";
 
 export const registerStudio = async (values: any, files: any) => {
@@ -8,7 +8,7 @@ export const registerStudio = async (values: any, files: any) => {
 
   // Campos de texto
   formData.append("name", values.name);
-  formData.append("studioType", values.studioType); // "grabacion" | "ensayo" | "produccion"
+  formData.append("studioType", values.studioType);"produccion"
   formData.append("city", values.city);
   formData.append("province", values.province);
   formData.append("address", values.address);
@@ -16,10 +16,8 @@ export const registerStudio = async (values: any, files: any) => {
   formData.append("openingTime", String(values.openingTime));
   formData.append("closingTime", String(values.closingTime));
 
-  // Servicios (array de strings)
-  values.services?.forEach((s: string) => formData.append("services[]", s));
 
-  // Equipamiento disponible (array de strings)
+  values.services?.forEach((s: string) => formData.append("services[]", s));
   values.availableEquipment?.forEach((e: string) => formData.append("availableEquipment[]", e));
 
   // Archivos
@@ -31,8 +29,6 @@ export const registerStudio = async (values: any, files: any) => {
       formData.append("photos", file);
     });
   }
-
-  // Request
   return await http.post("/owners/me/studio/files", formData, {
     headers: {
       Authorization: `Bearer ${token}`,
