+++ types/Rooms.ts
@@ -1,19 +1,23 @@
 export interface Availability {
-  [day: string]: {
-    start: string;
-    end: string;
-  };
+    [day: string]: {
+        start: string;
+        end: string;
+    };
 }
 
 export interface Room {
-  name: string;
-  type: string;
-  capacity: number;
-  size: number;
-  pricePerHour: number;
-  minHours: number;
-  description: string;
-  features: string[];
-  customEquipment?: string;
-  availability?: Availability;
+    id: string;
+    name: string;
+    type: string;
+    capacity: number;
+    pricePerHour: string;
+    size?: number;
+    minHours: number;
+    description: string;
+    features: string[];
+    customEquipment?: string;
+    availability?: Availability;
+    imageUrls?: string[] | null;
+    imagePublicIds?: string[] | null;
+    isActive?: boolean;
 }
